Reorganize GUI components into hierarchical structure
